FIX import de sanitizename et rename du fichier
