+++ apps/bot/src/domains/player/service.ts
@@ -1,8 +1,9 @@
 import type { Player } from '@one-piece/db';
 
+import { sanitizeName } from '../../shared/sanitize-name.js';
 import { findOrCreateShip } from '../ship/service.js';
 
-import { assertNameNotEmpty, assertNameWithinMaxLength, sanitizeName } from './name.js';
+import { assertNameNotEmpty, assertNameWithinMaxLength } from './name.js';
 import * as playerRepository from './repository.js';
 
 type FindOrCreateResult = { player: Player; created: boolean };
